feat: comprehensive update of backend services and frontend components - state at 2026-03-07

+++ backend-todo/src/main/java/com/loki/todo/dto/TodoDTO.java
@@ -30,7 +30,7 @@ public class TodoDTO {
     private LocalDateTime updatedAt;
     private LocalDateTime startedAt;
     private Double estimatedHours;
-    private Integer actualHours;
+    private Double actualHours;
     private Integer storyPoints;
 
     // User/Assignment fields
